Замена int на long

+++ src/Main.java
@@ -3,7 +3,7 @@ import java.util.Arrays;
 public class Main {
     public static void main(String[] args) {
 
-        int [] manager = {2,3,4};
+        long [] manager = {2,3,4};
 
         SalesManager salesManager= new SalesManager(manager);
         System.out.println(salesManager.max());
